requirements.txt
Flask==2.0.3
loguru==0.5.3
PyExecJS==1.5.1
requests==2.24.0
itsdangerous==2.0.1
gevent==21.12.0
gunicorn
